refactor graphql instrumentation for "NH" ...
... because we wont have instrumentation residing in the graphql gem

- remove code that checks version and make sure the newer one is used
- remove corresponding tests and test files
- supported versions now 1.8.0 - 2.0.+

NH-10331
