+++ test/noop/noop_test.rb
@@ -87,7 +87,7 @@ class NoopTest < Minitest::Test
   end
 
   def test_graphql_not_instrumented
-    refute GraphQL::Schema.plugins.find { |plugin| plugin == GraphQL::Tracing::AppOpticsTracing },
+    refute GraphQL::Schema.plugins.find { |plugin| plugin == GraphQL::Tracing::SolarWindsAPMTracing },
            'failed: This should be noop mode, but GraphQL is instrumented.'
   end
